feat(nvim): split files
