+++ .config/nvim/lua/plugins/configs/plantuml.lua
@@ -1,0 +1,4 @@
+return {
+	"aklt/plantuml-syntax",
+	event = { "BufReadPre" },
+}
